LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/move-current-target-expected.txt


FAIL Moving the current snap target should make the scroller resnap to it. assert_equals: expected 100 but got 0
FAIL Changing the layout of other elements should be able to cause resnapping to the target. assert_equals: expected 100 but got 0
FAIL Transforming the current snap target should make the scroller resnap to it. assert_equals: expected 100 but got 0
FAIL Applying two property changes that do not change the visual offset of the target should not change the scroll offset. assert_equals: expected 200 but got 0
